# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  services.AutoCompleteResponse:
    properties:
      suggestions:
        items:
          $ref: '#/definitions/services.AutoCompleteSuggestion'
        type: array
    type: object
  services.AutoCompleteSuggestion:
    properties:
      score:
        type: number
      text:
        type: string
      type:
        type: string
    type: object
  services.DateRange:
    properties:
      from:
        type: string
      to:
        type: string
    type: object
  services.SearchRequest:
    properties:
      author_id:
        type: string
      date_range:
        $ref: '#/definitions/services.DateRange'
      facets:
        items:
          type: string
        type: array
      filters:
        additionalProperties:
          type: string
        type: object
      from:
        type: integer
      fuzzy:
        type: boolean
      highlight:
        type: boolean
      privacy:
        type: string
      query:
        type: string
      size:
        type: integer
      sort_by:
        description: relevance, date, popularity
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        description: post, user, comment
        type: string
    type: object
  services.SearchResponse:
    properties:
      facets:
        additionalProperties: true
        type: object
      has_more:
        type: boolean
      max_score:
        type: number
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/services.SearchResult'
        type: array
      size:
        type: integer
      suggestions:
        items:
          type: string
        type: array
      took:
        description: Duration in milliseconds
        type: integer
      total:
        type: integer
    type: object
  services.SearchResult:
    properties:
      author_id:
        type: string
      author_name:
        type: string
      content:
        type: string
      created_at:
        type: string
      highlights:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      privacy:
        type: string
      score:
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive social media backend with posts, video calling, and
    file uploads
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media Backend API
  version: "2.0"
paths:
...
  /search:
    get:
      description: Search across posts, users, and comments with flexible filtering
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Content type filter (post, user, comment)
        in: query
        name: type
        type: string
      - description: Filter by author ID
        in: query
        name: author_id
        type: string
      - description: Privacy filter (public, friends, private)
        in: query
        name: privacy
        type: string
      - default: relevance
        description: Sort order (relevance, date, popularity)
        in: query
        name: sort_by
        type: string
      - default: 0
        description: Pagination offset
        in: query
        name: from
        type: integer
      - default: 20
        description: Number of results (1-100)
        in: query
        name: size
        type: integer
      - default: false
        description: Enable result highlighting
        in: query
        name: highlight
        type: boolean
      - default: false
        description: Enable fuzzy search
        in: query
        name: fuzzy
        type: boolean
      - collectionFormat: csv
        description: Filter by tags
        in: query
        items:
          type: string
        name: tags
        type: array
      - collectionFormat: csv
        description: Enable faceted search
        in: query
        items:
          type: string
        name: facets
        type: array
      - description: Start date filter (RFC3339 format)
        in: query
        name: from_date
        type: string
      - description: End date filter (RFC3339 format)
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Search failed
          schema:
            additionalProperties: true
            type: object
      summary: Global search
      tags:
      - Search
  /search/admin/documents/{id}:
    delete:
      description: Remove a specific document from the search index (admin only)
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Delete failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete search document
      tags:
      - Search
  /search/admin/reindex:
    post:
      description: Rebuild the entire search index (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Reindex successful
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Reindex failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reindex all content
      tags:
      - Search
  /search/admin/stats:
    get:
      description: Get statistics about the search index (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Search statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Stats failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get search statistics
      tags:
      - Search
  /search/advanced:
    post:
      consumes:
      - application/json
      description: Perform advanced search with complex JSON-based filtering
      parameters:
      - description: Advanced search parameters
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/services.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Search failed
          schema:
            additionalProperties: true
            type: object
      summary: Advanced search
      tags:
      - Search
  /search/autocomplete:
    get:
      description: Get search suggestions for partial queries
      parameters:
      - description: Partial search query
        in: query
        name: q
        required: true
        type: string
      - description: Content type filter (post, user, comment)
        in: query
        name: type
        type: string
      - default: 10
        description: Number of suggestions (1-50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search suggestions
          schema:
            $ref: '#/definitions/services.AutoCompleteResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Autocomplete failed
          schema:
            additionalProperties: true
            type: object
      summary: Auto-complete search
      tags:
      - Search
  /search/comments:
    get:
      description: Search for comments with filtering options
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 0
        description: Pagination offset
        in: query
        name: from
        type: integer
      - default: 20
        description: Number of results (1-100)
        in: query
        name: size
        type: integer
      - description: Filter by author ID
        in: query
        name: author_id
        type: string
      - default: false
        description: Enable result highlighting
        in: query
        name: highlight
        type: boolean
      - default: relevance
        description: Sort order (relevance, date)
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment search results
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "500":
          description: Search failed
          schema:
            additionalProperties: true
            type: object
      summary: Search comments
      tags:
      - Search
  /search/posts:
    get:
      description: Search for posts with specific filters
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 0
        description: Pagination offset
        in: query
        name: from
        type: integer
      - default: 20
        description: Number of results (1-100)
        in: query
        name: size
        type: integer
      - description: Filter by author ID
        in: query
        name: author_id
        type: string
      - description: Privacy filter (public, friends, private)
        in: query
        name: privacy
        type: string
      - collectionFormat: csv
        description: Filter by tags
        in: query
        items:
          type: string
        name: tags
        type: array
      - default: false
        description: Enable result highlighting
        in: query
        name: highlight
        type: boolean
      - default: relevance
        description: Sort order (relevance, date, popularity)
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post search results
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "500":
          description: Search failed
          schema:
            additionalProperties: true
            type: object
      summary: Search posts
      tags:
      - Search
  /search/trending-tags:
    get:
      description: Get currently trending hashtags based on post frequency
      parameters:
      - default: 10
        description: Number of tags to return (1-50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trending tags
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Trends failed
          schema:
            additionalProperties: true
            type: object
      summary: Get trending tags
      tags:
      - Search
  /search/users:
    get:
      description: Search for users by username, display name, or bio
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 0
        description: Pagination offset
        in: query
        name: from
        type: integer
      - default: 20
        description: Number of results (1-100)
        in: query
        name: size
        type: integer
      - default: false
        description: Enable result highlighting
        in: query
        name: highlight
        type: boolean
      - default: relevance
        description: Sort order (relevance, date)
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User search results
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "500":
          description: Search failed
          schema:
            additionalProperties: true
            type: object
      summary: Search users
      tags:
      - Search
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...